+++ src/Convo/Core/Adapters/Alexa/AlexaSkillPublisher.php
@@ -1200,9 +1200,7 @@ class AlexaSkillPublisher extends \Convo\Core\Publish\AbstractServicePublisher
 
         $config = $this->_getAmazonConfig();
 
-        $meta = $this->_convoServiceDataProvider->getServiceMeta(
-            $this->_user, $this->_serviceId
-        );
+        $meta = $this->_getServiceMeta();
 
         $owner = $this->_adminUserDataProvider->findUser($meta['owner']);
 
@@ -1242,11 +1240,12 @@ class AlexaSkillPublisher extends \Convo\Core\Publish\AbstractServicePublisher
 
 		$permissions = $config['permissions'] ?? [];
 
+        $public_name = $config['skill_preview_in_store']['public_name'] ?? $this->_invocationToName($meta['name']);
         $manifest->setGlobalEndpoint(
             $this->_serviceReleaseManager->getWebhookUrl(
                 $this->_user, $this->_serviceId, $this->getPlatformId()
             )
-        )->setName($locales, $config['skill_preview_in_store']['public_name'])
+        )->setName($locales, $public_name)
             ->setSummary($locales, $config['skill_preview_in_store']['one_sentence_description'])
             ->setDescription($locales, $config['skill_preview_in_store']['detailed_description'])
             ->setWhatsNew($locales, $config['skill_preview_in_store']['whats_new'])