+++ src/Convo/Pckg/Core/Elements/ElementCollection.php
@@ -2,10 +2,9 @@
 
 namespace Convo\Pckg\Core\Elements;
 
+use Convo\Core\Workflow\IConversationElement;
 
-use Convo\Core\Workflow\IOptionalElement;
-
-class ElementCollection extends \Convo\Core\Workflow\AbstractWorkflowContainerComponent implements \Convo\Core\Workflow\IConversationElement
+class ElementCollection extends \Convo\Core\Workflow\AbstractWorkflowContainerComponent implements IConversationElement
 {
 	
 	/**
@@ -27,9 +26,10 @@ class ElementCollection extends \Convo\Core\Workflow\AbstractWorkflowContainerCo
 	
 	public function read( \Convo\Core\Workflow\IConvoRequest $request, \Convo\Core\Workflow\IConvoResponse $response)
 	{
-		$this->_logger->debug('Reading ['.count( $this->_elements).']');
+		$this->_logger->info('Reading ['.count( $this->_elements).']');
 		
-		foreach ($this->getElements() as $element) {
+		foreach ($this->getElements() as $i=>$element) {
+		    $this->_logger->info('Reading element at index ['.$i.']');
 			/** @var $element \Convo\Core\Workflow\IConversationElement */
 			$element->read( $request, $response);
 		}
@@ -39,21 +39,7 @@ class ElementCollection extends \Convo\Core\Workflow\AbstractWorkflowContainerCo
 	 * @return \Convo\Core\Workflow\IConversationElement[]
 	 */
 	public function getElements() {
-	    $elements   =   $this->getService()->spreadElements( $this->_elements);
-	    $filtered   =   [];
-	    
-	    foreach ( $elements as $element) {
-	        if ( $element instanceof IOptionalElement) {
-	            /* @var IOptionalElement $element*/
-	            if ( $element->isEnabled()) {
-	                $filtered[] = $element;
-	            }
-	            continue;
-	        }
-	        $filtered[] = $element;
-	    }
-	    
-	    return $filtered;
+	    return $this->getService()->spreadElements( $this->_elements);
 	}
 	
 	// UTIL